requirements.txt
fastapi
uvicorn[standard]
python-multipart
dnspython
aiofiles
celery[redis]
redis
pandas
pydantic
python-dotenv
